requirements.txt
flask
openai
google-generativeai
anthropic
pyttsx3